Improve the flexibility of T type, KdTree.size and generate nearest_one_point o Relax the trait constraints on KdTree content type T and use T::default() o Make KdTree.size a usize instead of a T, simplifying interface and semantics o To allow use of the K-D Tree coordinates, provide nearest_one_point (in   addition to nearest_one), which returns the NearestNeighbour and also the   coordinate of the nearest K-D Tree point.

+++ src/fixed/query/within.rs
@@ -136,7 +136,7 @@ mod tests {
         content_to_add
             .iter()
             .for_each(|(point, content)| tree.add(point, *content));
-        assert_eq!(tree.size(), TREE_SIZE as u32);
+        assert_eq!(tree.size(), TREE_SIZE);
 
         let query_points: Vec<[Fxd; 4]> = (0..NUM_QUERIES)
             .map(|_| rand_data_fixed_u16_point::<U14, 4>())
